tartan_acpi::aml

Module name

Names of objects, arguments, and references

Structs

NameSeg

Four-character name segment, allowing underscores, uppercase letters, and digits (except at the beginning).

NameString

Fully qualified object path, either absolute or relative.

Enums

PathAnchor

Indicates whether a name is absolute or relative to the current or parent scope.

SimpleName

A named object or variable.

SuperName

A named object, variable, reference expression, or debug object.

Functions

to_path

Convert a list of segments in various formats into a path vector

Type Aliases

Target

Location to store the result of an operation