Struct tartan_acpi::aml::parse::state::ParserState

pub struct ParserState<'a> {
    pub data: &'a [u8],
    pub current_scope: Vec<NameSeg>,
    pub method_signatures: Vec<MethodSignature>,
}

Input data plus context needed to disambiguate grammar

Fields
data: &'a [u8]current_scope: Vec<NameSeg>method_signatures: Vec<MethodSignature>
Implementations

impl<'a> ParserState<'a>

pub fn new(data: &'a [u8]) -> Self

pub fn get_arg_count(&self, name: &NameString) -> Option<u8>

Get the argument count of the named method, if it is declared

pub fn in_scope<P, O, E>( self, new_scope: &NameString, parser: P, ) -> AMLParseResult<'a, O, E>
where P: Parser<ParserState<'a>, O, E>, E: AMLParseError<'a>,

Execute a parser in a new scope relative to the current scope, then restore the original scope.

Errors

Fails if the given scope is invalid when resolved against the current scope (because it falls outside the namespace root), or when the inner parser fails.

pub fn lift<P, O, E>( parser: P, ) -> impl FnMut(ParserState<'a>) -> AMLParseResult<'_, O, E>
where P: Parser<&'a [u8], O, E::From>, E: AMLParseError<'a>,

Wraps a byte-oriented parser to work with ParserState