Trait tartan_acpi::aml::parse::Parse

pub trait Parse<'a>
where
    Self: Sized,
{
    // Required method
    fn parse<E: AMLParseError<'a>>(
        i: ParserState<'a>
    ) -> AMLParseResult<'_, Self, E>;
}

An object that can be parsed from AML bytecode

Required Methods

fn parse<E: AMLParseError<'a>>( i: ParserState<'a> ) -> AMLParseResult<'_, Self, E>

Try to parse an object of this type from the given input and state.

Errors

Returns an error if the data cannot be parsed.

Object Safety

This trait is not object safe.