Enum tartan_acpi::aml::term::NamedObject

pub enum NamedObject<'a> {

    BankField {
        region_name: NameString,
        bank_name: NameString,
        bank_value: TermArg<'a>,
        flags: FieldFlags,
        elements: Vec<FieldElement<'a>>,
    },
    CreateBitField {
        source_buffer: TermArg<'a>,
        bit_index: TermArg<'a>,
        name: NameString,
    },
    CreateByteField {
        source_buffer: TermArg<'a>,
        byte_index: TermArg<'a>,
        name: NameString,
    },
    CreateDWordField {
        source_buffer: TermArg<'a>,
        byte_index: TermArg<'a>,
        name: NameString,
    },
    CreateField {
        source_buffer: TermArg<'a>,
        bit_index: TermArg<'a>,
        num_bits: TermArg<'a>,
        name: NameString,
    },
    CreateQWordField {
        source_buffer: TermArg<'a>,
        byte_index: TermArg<'a>,
        name: NameString,
    },
    CreateWordField {
        source_buffer: TermArg<'a>,
        byte_index: TermArg<'a>,
        name: NameString,
    },
    DataTableRegion {
        name: NameString,
        signature: TermArg<'a>,
        oem_id: TermArg<'a>,
        oem_table_id: TermArg<'a>,
    },
    Device {
        name: NameString,
        body: Vec<TermObject<'a>>,
    },
    Event(NameString),
    External {
        name: NameString,
        object_type: ObjectType,
        argument_count: u8,
    },
    Field {
        region_name: NameString,
        flags: FieldFlags,
        elements: Vec<FieldElement<'a>>,
    },
    IndexField {
        index_name: NameString,
        data_name: NameString,
        flags: FieldFlags,
        elements: Vec<FieldElement<'a>>,
    },
    Method {
        name: NameString,
        flags: MethodFlags,
        body: Vec<TermObject<'a>>,
    },
    Mutex {
        name: NameString,
        sync_level: u8,
    },
    OperationRegion {
        name: NameString,
        region_space: RegionSpace,
        offset: TermArg<'a>,
        length: TermArg<'a>,
    },
    PowerResource {
        name: NameString,
        system_level: u8,
        resource_order: u16,
        body: Vec<TermObject<'a>>,
    },
    Processor {
        name: NameString,
        id: u8,
        register_block_addr: u32,
        register_block_length: u8,
        body: Vec<TermObject<'a>>,
    },
    ThermalZone {
        name: NameString,
        body: Vec<TermObject<'a>>,
    },

}

Term that defines an object with a name.

Variants

BankField

Fields

region_name: NameString

bank_name: NameString

bank_value: TermArg<'a>

flags: FieldFlags

elements: Vec<FieldElement<'a>>

Declare fields that can only be accessed after writing to a bank selector

CreateBitField

Fields

source_buffer: TermArg<'a>

bit_index: TermArg<'a>

name: NameString

Declare a single-bit field within a buffer

CreateByteField

Fields

source_buffer: TermArg<'a>

byte_index: TermArg<'a>

name: NameString

Create a one-byte-wide field within a buffer

CreateDWordField

Fields

source_buffer: TermArg<'a>

byte_index: TermArg<'a>

name: NameString

Declare a four-byte-wide field within a buffer

CreateField

Fields

source_buffer: TermArg<'a>

bit_index: TermArg<'a>

num_bits: TermArg<'a>

name: NameString

Declare an arbitrary-width field within a buffer

CreateQWordField

Fields

source_buffer: TermArg<'a>

byte_index: TermArg<'a>

name: NameString

Declare an eight-byte-wide field within a buffer

CreateWordField

Fields

source_buffer: TermArg<'a>

byte_index: TermArg<'a>

name: NameString

Declare a two-byte-wide field within a buffer

DataTableRegion

Fields

name: NameString

signature: TermArg<'a>

oem_id: TermArg<'a>

oem_table_id: TermArg<'a>

Allow an ACPI table indexed in the XSDT to be used as a set of fields

Device

Fields

name: NameString

body: Vec<TermObject<'a>>

Declare a device and its associated fields, methods, and sub-devices

Event(NameString)

Declare a waitable synchronization object

External

Fields

name: NameString

object_type: ObjectType

argument_count: u8

Declare an object that is defined in another ACPI table

Field

Fields

region_name: NameString

flags: FieldFlags

elements: Vec<FieldElement<'a>>

Declare a group of fields

IndexField

Fields

index_name: NameString

data_name: NameString

flags: FieldFlags

elements: Vec<FieldElement<'a>>

Declare a group of field that must be accessed by writing to an index register and then reading/writing from a data register.

Method

Fields

name: NameString

flags: MethodFlags

body: Vec<TermObject<'a>>

Declare a control method

Mutex

Fields

name: NameString

sync_level: u8

Declare an acquirable mutex

OperationRegion

Fields

name: NameString

region_space: RegionSpace

offset: TermArg<'a>

length: TermArg<'a>

Declare an address space that can be used by fields

PowerResource

Fields

name: NameString

system_level: u8

resource_order: u16

body: Vec<TermObject<'a>>

Declare a power resource object

Processor

Fields

name: NameString

id: u8

register_block_addr: u32

register_block_length: u8

body: Vec<TermObject<'a>>

Declare a processor and associated register block

ThermalZone

Fields

name: NameString

body: Vec<TermObject<'a>>

Declare a thermal zone namespace