Module tartan_acpi::aml::data

Data resources

Structs

Buffer
Block of raw bytes, optionally initialized in whole or in part.

Package
Group of related data elements, optionally inititalized in whole or in part.

VarPackage
Group of related data elements, with a number of elements only known at runtime.

Enums

ComputationalData
Single value resolved at compile time.

DataObject
Data resolved at compile time, possibly grouped in a package.

DataRefObject

PackageElement
A name or reference that constitutes part of a Package (or VarPackage).