Function tartan_acpi::aml::parse::name::super_name_minus_invoke

pub fn super_name_minus_invoke<'a, E: AMLParseError<'a>>(
    i: ParserState<'a>
) -> AMLParseResult<'_, SuperName<'a>, E>

Parse a SuperName, but do not allow method invocations.

This is needed by some operators where invocations are not allowed, but bare NameString references are---including method references. The standard SuperName parser would try to parse it as an invocation first.