tartan_acpi::aml::parse

Trait Parse

pub trait Parse<'a>
where
    Self: Sized,
{
    // Required method
    fn parse<E: AMLParseError<'a>>(
        i: ParserState<'a>,
    ) -> AMLParseResult<'a, Self, E>;
}

An object that can be parsed from AML bytecode

Required Methods

fn parse<E: AMLParseError<'a>>( i: ParserState<'a>, ) -> AMLParseResult<'a, Self, E>

Try to parse an object of this type from the given input and state.

Errors

Returns an error if the data cannot be parsed.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.