Module tartan_acpi::aml::term

Top-level terms and opcodes.

Structs

FieldFlags

Rules for reading and writing a field.

MethodFlags

Information about how to call a method.

Enums

AccessAttrib

Additional information about how a field is accessed.

AccessType

Width used to access a field.

ExpressionOpcode

Terminal operation that evaluates to a value or reference.

FieldElement

A substructure of a field.

MatchOpcode

Type of comparison used for a branch in an ExpressionOpcode::Match expression.

NameSpaceModifier

Term that attaches a name to its argument.

NamedObject

Term that defines an object with a name.

ObjectType

Type ID returned by ExpressionOpcode::ObjectType.

ReferenceExpressionOpcode

Terminal operation that evaluates to a reference.

RegionSpace

Address space used to access a field.

StatementOpcode

Terminal operation that does not evaluate to a value, a.k.a. “Type 1 Opcode.”

TermArg

Term that resolves to a value.

TermObject

Top-level, most general term type where the value (if any) is discarded.

UpdateRule

Indicates what to do with unused bits when writing a field.