Enum tartan_acpi::aml::term::NamedObject

pub enum NamedObject<'a> {
    BankField {
        region_name: NameString,
        bank_name: NameString,
        bank_value: TermArg<'a>,
        flags: FieldFlags,
        elements: Vec<FieldElement<'a>>,
    },
    CreateBitField {
        source_buffer: TermArg<'a>,
        bit_index: TermArg<'a>,
        name: NameString,
    },
    CreateByteField {
        source_buffer: TermArg<'a>,
        byte_index: TermArg<'a>,
        name: NameString,
    },
    CreateDWordField {
        source_buffer: TermArg<'a>,
        byte_index: TermArg<'a>,
        name: NameString,
    },
    CreateField {
        source_buffer: TermArg<'a>,
        bit_index: TermArg<'a>,
        num_bits: TermArg<'a>,
        name: NameString,
    },
    CreateQWordField {
        source_buffer: TermArg<'a>,
        byte_index: TermArg<'a>,
        name: NameString,
    },
    CreateWordField {
        source_buffer: TermArg<'a>,
        byte_index: TermArg<'a>,
        name: NameString,
    },
    DataTableRegion {
        name: NameString,
        signature: TermArg<'a>,
        oem_id: TermArg<'a>,
        oem_table_id: TermArg<'a>,
    },
    Device {
        name: NameString,
        body: Vec<TermObject<'a>>,
    },
    Event(NameString),
    External {
        name: NameString,
        object_type: ObjectType,
        argument_count: u8,
    },
    Field {
        region_name: NameString,
        flags: FieldFlags,
        elements: Vec<FieldElement<'a>>,
    },
    IndexField {
        index_name: NameString,
        data_name: NameString,
        flags: FieldFlags,
        elements: Vec<FieldElement<'a>>,
    },
    Method {
        name: NameString,
        flags: MethodFlags,
        body: Vec<TermObject<'a>>,
    },
    Mutex {
        name: NameString,
        sync_level: u8,
    },
    OperationRegion {
        name: NameString,
        region_space: RegionSpace,
        offset: TermArg<'a>,
        length: TermArg<'a>,
    },
    PowerResource {
        name: NameString,
        system_level: u8,
        resource_order: u16,
        body: Vec<TermObject<'a>>,
    },
    Processor {
        name: NameString,
        id: u8,
        register_block_addr: u32,
        register_block_length: u8,
        body: Vec<TermObject<'a>>,
    },
    ThermalZone {
        name: NameString,
        body: Vec<TermObject<'a>>,
    },
}

Term that defines an object with a name.

Variants

BankField

Declare fields that can only be accessed after writing to a bank selector

Fields of BankField

region_name: NameStringbank_name: NameStringbank_value: TermArg<'a>flags: FieldFlagselements: Vec<FieldElement<'a>>

CreateBitField

Declare a single-bit field within a buffer

Fields of CreateBitField

source_buffer: TermArg<'a>bit_index: TermArg<'a>name: NameString

CreateByteField

Create a one-byte-wide field within a buffer

Fields of CreateByteField

source_buffer: TermArg<'a>byte_index: TermArg<'a>name: NameString

CreateDWordField

Declare a four-byte-wide field within a buffer

Fields of CreateDWordField

source_buffer: TermArg<'a>byte_index: TermArg<'a>name: NameString

CreateField

Declare an arbitrary-width field within a buffer

Fields of CreateField

source_buffer: TermArg<'a>bit_index: TermArg<'a>num_bits: TermArg<'a>name: NameString

CreateQWordField

Declare an eight-byte-wide field within a buffer

Fields of CreateQWordField

source_buffer: TermArg<'a>byte_index: TermArg<'a>name: NameString

CreateWordField

Declare a two-byte-wide field within a buffer

Fields of CreateWordField

source_buffer: TermArg<'a>byte_index: TermArg<'a>name: NameString

DataTableRegion

Allow an ACPI table indexed in the XSDT to be used as a set of fields

Fields of DataTableRegion

name: NameStringsignature: TermArg<'a>oem_id: TermArg<'a>oem_table_id: TermArg<'a>

Device

Declare a device and its associated fields, methods, and sub-devices

Fields of Device

name: NameStringbody: Vec<TermObject<'a>>

Event(NameString)

Declare a waitable synchronization object

External

Declare an object that is defined in another ACPI table

Fields of External

name: NameStringobject_type: ObjectTypeargument_count: u8

Field

Declare a group of fields

Fields of Field

region_name: NameStringflags: FieldFlagselements: Vec<FieldElement<'a>>

IndexField

Declare a group of field that must be accessed by writing to an index register and then reading/writing from a data register.

Fields of IndexField

index_name: NameStringdata_name: NameStringflags: FieldFlagselements: Vec<FieldElement<'a>>

Method

Declare a control method

Fields of Method

name: NameStringflags: MethodFlagsbody: Vec<TermObject<'a>>

Mutex

Declare an acquirable mutex

Fields of Mutex

name: NameStringsync_level: u8

OperationRegion

Declare an address space that can be used by fields

Fields of OperationRegion

name: NameStringregion_space: RegionSpaceoffset: TermArg<'a>length: TermArg<'a>

PowerResource

Declare a power resource object

Fields of PowerResource

name: NameStringsystem_level: u8resource_order: u16body: Vec<TermObject<'a>>

Processor

Declare a processor and associated register block

Fields of Processor

name: NameStringid: u8register_block_addr: u32register_block_length: u8body: Vec<TermObject<'a>>

ThermalZone

Declare a thermal zone namespace

Fields of ThermalZone

name: NameStringbody: Vec<TermObject<'a>>