Enum tartan_acpi::aml::term::StatementOpcode

pub enum StatementOpcode<'a> {
    Break,
    BreakPoint,
    Continue,
    Fatal {
        fatal_type: u8,
        code: u32,
        arg: TermArg<'a>,
    },
    If {
        predicate: TermArg<'a>,
        if_true: Vec<TermObject<'a>>,
        if_false: Vec<TermObject<'a>>,
    },
    Load {
        name: NameString,
        definition_block_handle: SuperName<'a>,
    },
    NoOp,
    Notify {
        device_or_zone: SuperName<'a>,
        value: TermArg<'a>,
    },
    Release {
        mutex: SuperName<'a>,
    },
    Reset {
        event: SuperName<'a>,
    },
    Return(TermArg<'a>),
    Signal {
        event: SuperName<'a>,
    },
    Sleep {
        milliseconds: TermArg<'a>,
    },
    Stall {
        microseconds: TermArg<'a>,
    },
    While {
        predicate: TermArg<'a>,
        body: Vec<TermObject<'a>>,
    },
}

Terminal operation that does not evaluate to a value, a.k.a. "Type 1 Opcode."

Variants

Break

Break from a loop

BreakPoint

Trigger a debugger breakpoint

Continue

Continue to the next loop iteration

Fatal

Signal an error that requires a system shutdown

Fields of Fatal

fatal_type: u8code: u32arg: TermArg<'a>

If

Branch on a predicate

Fields of If

predicate: TermArg<'a>if_true: Vec<TermObject<'a>>if_false: Vec<TermObject<'a>>

Load

Load a dynamically-generated SSDT from a field, region, or buffer

Fields of Load

name: NameStringdefinition_block_handle: SuperName<'a>

NoOp

Do nothing

Notify

Send a signal value to a device/processor/zone

Fields of Notify

device_or_zone: SuperName<'a>value: TermArg<'a>

Release

Release a held mutex

Fields of Release

mutex: SuperName<'a>

Reset

Clear the signalled state of an event object

Fields of Reset

event: SuperName<'a>

Return(TermArg<'a>)

Exit the current method and yield the given value to the caller

Signal

Signal to one thread waiting on the event

Fields of Signal

event: SuperName<'a>

Sleep

Delay for at least the given milliseconds, releasing the processor

Fields of Sleep

milliseconds: TermArg<'a>

Stall

Delay for at least the given microseconds, but do not release the processor

Fields of Stall

microseconds: TermArg<'a>

While

Execute a series of statements as long as the predicate is true

Fields of While

predicate: TermArg<'a>body: Vec<TermObject<'a>>